Add files via upload
When you create a repository on GitHub, you have the option to select a template from a list of options. This will pre-populate your repository with the files and folders required for that particular template. Alternatively, you can create your own template by creating a repository with the required files and folders, and then using it as a starting point for your future projects.